E5b support

+++ src/decode/decode_bds_b2.c
@@ -52,7 +52,7 @@ static decoder_interface_list_element_t list_element_bds_b2 = {
     .interface = &decoder_interface_bds_b2, .next = NULL};
 
 void decode_bds_b2_register(void) {
-  for (u32 i = 0; i < ARRAY_SIZE(bds_b2_decoders); i++) {
+  for (u16 i = 0; i < ARRAY_SIZE(bds_b2_decoders); i++) {
     bds_b2_decoders[i].active = false;
     bds_b2_decoders[i].data = &bds_b2_decoder_data[i];
   }
