+++ src/decode/decode_qzss_l1ca.c
@@ -53,7 +53,7 @@ static decoder_interface_list_element_t list_element_qzss_l1ca = {
     .interface = &decoder_interface_qzss_l1ca, .next = NULL};
 
 void decode_qzss_l1ca_register(void) {
-  for (u32 i = 0; i < ARRAY_SIZE(qzss_l1ca_decoders); i++) {
+  for (u16 i = 0; i < ARRAY_SIZE(qzss_l1ca_decoders); i++) {
     qzss_l1ca_decoders[i].active = false;
     qzss_l1ca_decoders[i].data = &qzss_l1ca_decoder_data[i];
   }
